残りの全ての型エラーを修正
- my_page_timetable.dartのnum型をdoubleに変換
- firebase_get_kadai.dartの型キャストを追加
- notification.dartの型キャストを追加
- supplementary_lecture_sorting.dartの型エラーを修正

+++ lib/feature/timetable/widget/my_page_timetable.dart
@@ -252,7 +252,7 @@ class MyPageTimetable extends ConsumerWidget {
     final buttonPosition = (DateTime.now().weekday - 1) * (buttonSize + buttonPadding);
     final initialScrollOffset =
         (buttonPosition > deviceCenter) ? buttonPosition - deviceCenter : 0;
-    final controller = ScrollController(initialScrollOffset: initialScrollOffset);
+    final controller = ScrollController(initialScrollOffset: initialScrollOffset.toDouble());
     return Consumer(
       builder: (context, ref, child) {
         ref.watch(saveTimetableProvider);
